First commit including a lot of work
This commit includes mostly what I've done for the front page and
navigation buttons. I was super inspired by all of the stuff over at
codrops, so I tried my best to at least imitate some of the techniques
they used over there.

The background is an image that has been stretched across the screen to
fit. and will remain static. The image was found on a free stock photo
site.

As the page loads, text fades in, first the logo and then the nav
buttons.

The nav buttons are by far the toughest thing I've done here so far.
They employ a variety of css3 animations as well as jquery to manipulate
the page and their appearance. They're dynamic in that after clicked,
they shrink and move to the bottom of the screen to make way for
content.

Todo: I need to start coming up with dummy content for the three pages.
All of them I'm going to try and make dynamic so that anyone with an
account (anyone who's been given an account by me in the family), will
be able to post news, photos, and register on the calendar. These will
be the main features of the website and why I decided to make it in the
first place.