Calendar Page half implemented
The site now feature's a half working Calendar. The calendar uses php to
determine which month it is and then assign's the dates to the
corresponding days of the week dynamically.

TODO: Implement a system to allow users to add themselves to the
calendar. This will most likely be accomplished by creating some kind of
database. Each time the calendar loads, the corresponding entries for
that month will be loaded.

Also, as of right now the only month visible is the current month. I'd
like to ideally implement some kind of paging system with nice animated
transitions that allow you to view the previous and next months.