Holy Crap it works: Clickable Calendar
The calendar now has clickable dates which opens a modal that, in the
future, will allow users to leave their name down.

The best part of the commit. I took on the daunting task of figuring out
how to get my website to transition back to the original landing page.
This ended up being extremely problematic, but I ended up figuring it
out after hours and hours of work. I'm going to bed. Tomorrow I'll wake
up and be rewarded with actually being able to see the cabin itself and
hopefully be able to take some great pictures to upload to the site.
Can't wait.