Backend for calendar implemented
The backend for the calendar is finally here. Heroku provided a really
easy to set up and use database.  The integration of the database was
another story. My code is incredibly messy and not very well thought
out. I'm at the stage in the process where I'm realizing this and
considering rewriting how the calendar works. For now, I'm happy with
it, but shitty code like this won't be getting me any jobs anytime soon.

What works now: Users can now leave reservations by clicking on a date
and leaving their name. After clicking submit they're redirected to the
index page. I realize this isn't ideal and I'm hoping I can find a way
to simply redirect back to the calendar itself with the updates
reservation list. After leaving a reservation users can also delete it
by clicking on the date and hitting cancel reservation. It's this basic,
simple functionality that I wanted to achieve so that my family would be
encouraged to use it.

What doesn't work: Users can only reserve dates for the current month.
This is obviously not ideal and the next step will be to implement
something that allows users to cycle through the months and leave
reservations further in the future. I'm sure I can come up with
something.

TODO: Calendar month cycling, better page redirecting, better photo
album slideshows, and user slideshows.